requirements.txt
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
openai==0.27.2
python-dotenv==1.0.0
Pillow==9.5.0